cijfer kan nu ingevoerd worden

+++ app/src/main/java/com/hsl/imtpmd/imtpmd/Hoofdfase1Fragment.java
@@ -94,7 +94,7 @@ public class Hoofdfase1Fragment extends Fragment {
                 UserVerplichtvakModel vak = verplichtvakModels.get(position);
                 b.putString("code", vak.getVerplichtvak().getCode());
                 b.putString("naam", vak.getVerplichtvak().getNaam());
-                b.putInt("cijfer", vak.getCijfer());
+                b.putDouble("cijfer", vak.getCijfer());
                 b.putBoolean("behaald", vak.getBehaald());
                 i.putExtras(b);
                 startActivity(i);
@@ -115,7 +115,7 @@ public class Hoofdfase1Fragment extends Fragment {
                 UserKeuzevakModel vak = keuzevakModels.get(position);
                 b.putString("code", vak.getKeuzevak().getCode());
                 b.putString("naam", vak.getKeuzevak().getNaam());
-                b.putInt("cijfer", vak.getCijfer());
+                b.putDouble("cijfer", vak.getCijfer());
                 b.putBoolean("behaald", vak.getBehaald());
                 i.putExtras(b);
                 startActivity(i);
@@ -137,7 +137,7 @@ public class Hoofdfase1Fragment extends Fragment {
                 UserSpecialisatievakModel vak = specialisatievakModels.get(position);
                 b.putString("code", vak.getSpecialisatievakModel().getCode());
                 b.putString("naam", vak.getSpecialisatievakModel().getNaam());
-                b.putInt("cijfer", vak.getCijfer());
+                b.putDouble("cijfer", vak.getCijfer());
                 b.putBoolean("behaald", vak.getBehaald());
                 i.putExtras(b);
                 startActivity(i);
